benchmark/requirements.txt
psycopg2-binary
snowflake-connector-python
colorama
